fix: add check-db command
